clean up folders, seperate styles from jsx
