+++ package.json
@@ -51,6 +51,7 @@
   "dependencies": {
     "axios": "^0.19.2",
     "formik": "^2.1.4",
+    "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-loading": "^2.0.3",
