+++ src/App.js
@@ -2,11 +2,10 @@ import React from 'react'
 import styled from 'styled-components'
 import { neutrals } from './utils'
 import { SignupPage } from './components/pages/Signup'
-
 const Div = styled.div`
-  background-color: ${neutrals.lightgrey100};
+  background-color: ${neutrals.white};
   width: 100%;
-  height: 100vh;
+  height: 100%;
 `
 
 function App() {
