+++ src/components/organisms/styled/index.js
@@ -1,0 +1,2 @@
+export * from './styledFooter'
+export * from './styledForms'
