--- src/components/pages/Signup.js
@@ -1,15 +1,0 @@
-import React from 'react'
-import styled from 'styled-components'
-import { SignupForm, Nav, Footer } from '../organisms'
-
-const Div = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`
-
-export const SignupPage = () => (
-  <Div>
-    <Footer />
-  </Div>
-)
